refactor: split ChecklistRunner into runner + QuestionInputs
Extracts all question input components (CheckboxInput, NumberInput,
TemperatureSliderInput, TextInput, MultipleChoiceInput, DateTimeInput,
InstructionBlock, MediaInput, QuestionInput) into QuestionInputs.tsx.
ChecklistRunner.tsx now contains only accordion/navigation/draft logic.
No behaviour change — pure file split to reduce per-session context load.
